move constants out of shared folder
