+++ packages/components/src/index.js
@@ -1,3 +1,4 @@
 export * from './Loc';
 export * from './make-localized-element';
 export * from './pseudolocalize';
+export * from './constants';
