+++ packages/extraction/package.json
@@ -70,6 +70,7 @@
   "dependencies": {
     "@babel/parser": "^7.1.0",
     "babylon-walk": "^1.0.2",
+    "fluent-react-components": "*",
     "fluent-syntax": "^0.8.1",
     "glob": "^7.1.3",
     "jsx-ast-utils": "^2.0.1",
