+++ packages/extraction/src/ast-helper.js
@@ -1,5 +1,6 @@
+import { STANDARD_ELEMENT_TYPES } from 'fluent-react-components';
 import { getProp, elementType as _elementType, hasProp } from 'jsx-ast-utils';
-import { AST_NODE_TYPES, FLUENT_ATTRS, STANDARD_ELEMENT_TYPES } from '../../../shared/constants';
+import { AST_NODE_TYPES, FLUENT_ATTRS } from './constants';
 import { defaultShorthandName } from './defaults';
 import { formatRule, pullLocalizedDOMAttributes, formatMessage } from './format-helper';
