+++ packages/extraction/src/constants.js
@@ -1,0 +1,14 @@
+export const AST_NODE_TYPES = {
+  JSXElement: 'JSXElement',
+  JSXText: 'JSXText',
+  JSXExpressionContainer: 'JSXExpressionContainer',
+  JSXEmptyExpression: 'JSXEmptyExpression',
+  StringLiteral: 'StringLiteral'
+};
+
+export const FLUENT_ATTRS = {
+  attrs: 'attrs',
+  l10nId: 'l10nId',
+  l10nVars: 'l10nVars',
+  l10nJSX: 'l10nJSX'
+};
